requirements-prod.txt
# Core Flask dependencies
Flask
Flask-SQLAlchemy
Flask-Session
Flask-CORS
gunicorn

# Database
psycopg2-binary
redis

# AI and ML
groq
google-generativeai
mem0ai
qdrant-client
sentence-transformers

# LangChain dependencies
langchain
langchain-core
langchain-community
langchain-groq
langchain-openai
langchain-qdrant

# LangGraph and Agent Memory Components
langgraph
langgraph-checkpoint
langgraph-checkpoint-redis
redisvl
ulid-py

# Utilities
python-dotenv
requests
numpy
pandas
Pillow

# Email (using built-in smtplib)

# Security
bcrypt
PyJWT

# Monitoring and logging 
coloredlogs

# Data processing
python-dateutil
pytz

# HTTP and networking
urllib3
certifi

# JSON and serialization
orjson

# Environment and configuration
pydantic
pydantic-settings

# Cache and session management
cachelib

# System monitoring
psutil

# Unique identifiers
ulid-py
